重构addon sendMsg
